requirements.txt
python-telegram-bot==21.9
openai>=1.40.0
python-dotenv>=1.0.1
fastapi==0.112.0
uvicorn[standard]==0.30.5
SQLAlchemy==2.0.32
pydantic==2.8.2
pandas==2.2.2
streamlit==1.37.1
python-multipart==0.0.9
requests